Quick refactor to rsx crate with some tests
