+++ packages/rsx/.vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "rust-analyzer.check.workspace": false,
+    // "rust-analyzer.check.extraArgs": [
+    //     "--features",
+    //     "hot_reload"
+    // ],
+    "rust-analyzer.cargo.features": "all"
+}
