+++ packages/rsx/Cargo.toml
@@ -23,7 +23,7 @@ krates = { version = "0.16.6", optional = true }
 tracing = { workspace = true }
 
 [features]
-default = ["html"]
+default = ["html", "hot_reload"]
 hot_reload = ["krates", "internment", "dioxus-core"]
 serde = ["dep:serde"]
 html = []
