+++ packages/rsx/tests/valid/for_.new.rsx
@@ -1,0 +1,9 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        for items in vec![1, 2, 3] {
+            div { "asasddasdasd" }
+        }
+    }
+}
