+++ packages/rsx/tests/valid/for_.old.rsx
@@ -1,0 +1,10 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        for items in vec![1, 2, 3] {
+            div { "123" }
+            div { "asasddasdasd" }
+        }
+    }
+}
