+++ packages/rsx/tests/valid/if_.new.rsx
@@ -1,0 +1,10 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        if cond() {
+            div { "123" }
+            div { "asasddasdasd" }
+        }
+    }
+}
