+++ packages/rsx/tests/valid/if_.old.rsx
@@ -1,0 +1,9 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        if cond() {
+            div { "asasddasdasd" }
+        }
+    }
+}
