+++ packages/rsx/tests/valid/let_.new.rsx
