+++ packages/rsx/tests/valid/let_.old.rsx
