+++ packages/rsx/tests/valid/nested.new.rsx
@@ -1,0 +1,10 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        ForLoop {
+            div { "123" }
+            div { "asasddasdasd" }
+        }
+    }
+}
