+++ packages/rsx/tests/valid/nested.old.rsx
@@ -1,0 +1,9 @@
+use dioxus::prelude::*;
+
+pub fn CoolChild() -> Element {
+    rsx! {
+        ForLoop {
+            div { "asasddasdasd" }
+        }
+    }
+}
